refactor: event upload logic and introduce EventUpload class
