+++ core/src/main/kotlin/com/rudderstack/sdk/kotlin/core/internals/queue/EventQueue.kt
@@ -3,64 +3,41 @@ package com.rudderstack.sdk.kotlin.core.internals.queue
 import com.rudderstack.sdk.kotlin.core.Analytics
 import com.rudderstack.sdk.kotlin.core.internals.logger.LoggerAnalytics
 import com.rudderstack.sdk.kotlin.core.internals.models.Event
-import com.rudderstack.sdk.kotlin.core.internals.network.HttpClient
-import com.rudderstack.sdk.kotlin.core.internals.network.HttpClientImpl
 import com.rudderstack.sdk.kotlin.core.internals.policies.FlushPoliciesFacade
 import com.rudderstack.sdk.kotlin.core.internals.storage.StorageKeys
-import com.rudderstack.sdk.kotlin.core.internals.utils.JsonSentAtUpdater
-import com.rudderstack.sdk.kotlin.core.internals.utils.Result
 import com.rudderstack.sdk.kotlin.core.internals.utils.empty
-import com.rudderstack.sdk.kotlin.core.internals.utils.encodeToBase64
 import com.rudderstack.sdk.kotlin.core.internals.utils.encodeToString
-import com.rudderstack.sdk.kotlin.core.internals.utils.generateUUID
 import com.rudderstack.sdk.kotlin.core.internals.utils.isSourceEnabled
-import com.rudderstack.sdk.kotlin.core.internals.utils.parseFilePaths
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.channels.Channel.Factory.UNLIMITED
-import kotlinx.coroutines.channels.consumeEach
-import kotlinx.coroutines.ensureActive
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import org.jetbrains.annotations.VisibleForTesting
-import java.io.File
-import java.io.FileNotFoundException
 
 internal const val UPLOAD_SIG = "#!upload"
-private const val BATCH_ENDPOINT = "/v1/batch"
-private val ANONYMOUS_ID_REGEX = """"anonymousId"\s*:\s*"([^"]+)"""".toRegex()
 
 @OptIn(DelicateCoroutinesApi::class)
 internal class EventQueue(
     private val analytics: Analytics,
     private var flushPoliciesFacade: FlushPoliciesFacade = FlushPoliciesFacade(analytics.configuration.flushPolicies),
-    private val jsonSentAtUpdater: JsonSentAtUpdater = JsonSentAtUpdater(),
-    private val httpClientFactory: HttpClient = with(analytics.configuration) {
-        return@with HttpClientImpl.createPostHttpClient(
-            baseUrl = dataPlaneUrl,
-            endPoint = BATCH_ENDPOINT,
-            authHeaderString = writeKey.encodeToBase64(),
-            isGZIPEnabled = gzipEnabled,
-            anonymousIdHeaderString = analytics.anonymousId ?: String.empty()
-        )
-    }
+    private val eventUpload: EventUpload = EventUpload(
+        analytics = analytics,
+    ),
 ) {
 
     private var running: Boolean
     private var writeChannel: Channel<QueueMessage>
-    private var uploadChannel: Channel<String>
-    private val storage get() = analytics.storage
+    private val storage
+        get() = analytics.storage
     private val flushSignal = QueueMessage(QueueMessage.QueueMessageType.FLUSH_SIGNAL)
     private var lastEventAnonymousId = storage.readString(
         StorageKeys.LAST_EVENT_ANONYMOUS_ID,
         analytics.anonymousId ?: String.empty()
     )
-    private var lastBatchAnonymousId = String.empty()
 
     init {
         running = false
         writeChannel = Channel(UNLIMITED)
-        uploadChannel = Channel(UNLIMITED)
     }
 
     internal fun put(event: Event) {
@@ -73,11 +50,11 @@ internal class EventQueue(
 
         if (writeChannel.isClosedForSend || writeChannel.isClosedForReceive) {
             writeChannel = Channel(UNLIMITED)
-            uploadChannel = Channel(UNLIMITED)
         }
+        eventUpload.start()
+
         observeConfigAndUpdateSchedule()
         write()
-        upload()
     }
 
     private fun observeConfigAndUpdateSchedule() {
@@ -102,7 +79,7 @@ internal class EventQueue(
         if (!running) return
         running = false
 
-        uploadChannel.cancel()
+        eventUpload.cancel()
         writeChannel.close()
 
         flushPoliciesFacade.cancelSchedule()
@@ -133,52 +110,8 @@ internal class EventQueue(
             }
 
             if ((isFlushSignal || flushPoliciesFacade.shouldFlush()) && analytics.isSourceEnabled()) {
-                uploadChannel.trySend(UPLOAD_SIG)
+                eventUpload.flush()
                 flushPoliciesFacade.reset()
-            }
-        }
-    }
-
-    @Suppress("TooGenericExceptionCaught")
-    private fun upload() = analytics.analyticsScope.launch(analytics.networkDispatcher) {
-        uploadChannel.consumeEach {
-            LoggerAnalytics.debug("performing flush")
-            withContext(analytics.storageDispatcher) {
-                storage.rollover()
-            }
-            val fileUrlList = storage.readString(StorageKeys.EVENT, String.empty()).parseFilePaths()
-            for (filePath in fileUrlList) {
-                val file = File(filePath)
-                if (!doesFileExist(file)) continue
-                // ensureActive is at this position so that this coroutine can be cancelled - but any uploaded event MUST be cleared from storage.
-                ensureActive()
-                var shouldCleanup = false
-                try {
-                    val batchPayload = jsonSentAtUpdater.updateSentAt(readFileAsString(filePath))
-
-                    checkAndUpdateBatchRequestHeader(batchPayload)
-                    LoggerAnalytics.debug("Batch Payload: $batchPayload")
-                    when (val result: Result<String, Exception> = httpClientFactory.sendData(batchPayload)) {
-                        is Result.Success -> {
-                            LoggerAnalytics.debug("Event uploaded successfully. Server response: ${result.response}")
-                            shouldCleanup = true
-                        }
-
-                        is Result.Failure -> {
-                            LoggerAnalytics.debug("Error when uploading event due to ${result.status} ${result.error}")
-                        }
-                    }
-                } catch (e: FileNotFoundException) {
-                    LoggerAnalytics.error("Message storage file not found", e)
-                } catch (e: Exception) {
-                    LoggerAnalytics.error("Error when uploading event", e)
-                }
-
-                if (shouldCleanup) {
-                    storage.remove(file.path).let {
-                        LoggerAnalytics.debug("Removed file: $filePath")
-                    }
-                }
             }
         }
     }
@@ -193,30 +126,6 @@ internal class EventQueue(
                 storage.write(StorageKeys.LAST_EVENT_ANONYMOUS_ID, lastEventAnonymousId)
             }
         }
-    }
-
-    private fun checkAndUpdateBatchRequestHeader(batchPayload: String) {
-        val currentBatchAnonymousId = getAnonymousIdFromBatch(batchPayload)
-        if (lastBatchAnonymousId != currentBatchAnonymousId) {
-            httpClientFactory.updateAnonymousIdHeaderString(currentBatchAnonymousId.encodeToBase64())
-            lastBatchAnonymousId = currentBatchAnonymousId
-        }
-    }
-
-    @VisibleForTesting
-    internal fun getAnonymousIdFromBatch(batchPayload: String): String {
-        return ANONYMOUS_ID_REGEX.find(batchPayload)?.groupValues?.get(1) ?: run {
-            LoggerAnalytics.error("Fetched empty anonymousId from batch payload, falling back to random UUID.")
-            generateUUID()
-        }
-    }
-
-    @VisibleForTesting
-    fun doesFileExist(file: File) = file.exists()
-
-    @VisibleForTesting
-    fun readFileAsString(filePath: String): String {
-        return File(filePath).readText()
     }
 }
