+++ core/src/test/kotlin/com/rudderstack/sdk/kotlin/core/internals/queue/EventQueueTest.kt
@@ -5,17 +5,11 @@ import com.rudderstack.sdk.kotlin.core.internals.logger.KotlinLogger
 import com.rudderstack.sdk.kotlin.core.internals.models.Event
 import com.rudderstack.sdk.kotlin.core.internals.models.SourceConfig
 import com.rudderstack.sdk.kotlin.core.internals.models.provider.provideEvent
-import com.rudderstack.sdk.kotlin.core.internals.network.ErrorStatus
-import com.rudderstack.sdk.kotlin.core.internals.network.HttpClient
-import com.rudderstack.sdk.kotlin.core.internals.utils.Result
 import com.rudderstack.sdk.kotlin.core.internals.policies.FlushPoliciesFacade
 import com.rudderstack.sdk.kotlin.core.internals.statemanagement.State
 import com.rudderstack.sdk.kotlin.core.internals.storage.Storage
 import com.rudderstack.sdk.kotlin.core.internals.storage.StorageKeys
-import com.rudderstack.sdk.kotlin.core.internals.utils.empty
-import com.rudderstack.sdk.kotlin.core.internals.utils.encodeToBase64
 import com.rudderstack.sdk.kotlin.core.internals.utils.encodeToString
-import com.rudderstack.sdk.kotlin.core.internals.utils.generateUUID
 import com.rudderstack.sdk.kotlin.core.mockAnalytics
 import com.rudderstack.sdk.kotlin.core.setupLogger
 import io.mockk.MockKAnnotations
@@ -25,7 +19,6 @@ import io.mockk.every
 import io.mockk.impl.annotations.MockK
 import io.mockk.just
 import io.mockk.mockk
-import io.mockk.mockkStatic
 import io.mockk.runs
 import io.mockk.spyk
 import io.mockk.unmockkAll
@@ -42,20 +35,12 @@ import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.Arguments
-import org.junit.jupiter.params.provider.MethodSource
-import java.io.FileNotFoundException
-import java.io.IOException
 
 @OptIn(ExperimentalCoroutinesApi::class)
 class EventQueueTest {
 
     @MockK
     private lateinit var mockStorage: Storage
-
-    @MockK
-    private lateinit var mockHttpClient: HttpClient
 
     @MockK
     private lateinit var mockKotlinLogger: KotlinLogger
@@ -92,7 +77,6 @@ class EventQueueTest {
         eventQueue = spyk(
             EventQueue(
                 analytics = mockAnalytics,
-                httpClientFactory = mockHttpClient,
                 flushPoliciesFacade = mockFlushPoliciesFacade,
             )
         )
@@ -140,246 +124,6 @@ class EventQueueTest {
 
         val result = eventQueue.stringifyBaseEvent(event)
         assertEquals(jsonString, result)
-    }
-
-    @Test
-    fun `test readFileAsString reads file correctly`() {
-        val filePath = "test_file_path"
-        val fileContent = "test content"
-
-        every { eventQueue.readFileAsString(filePath) } returns fileContent
-
-        val result = eventQueue.readFileAsString(filePath)
-
-        assertEquals(fileContent, result)
-    }
-
-    @Test
-    fun `given multiple batch is ready to be sent to the server and server returns success, when flush is called, then all the batches are sent to the server and removed from the storage`() =
-        runTest {
-            val storage = mockAnalytics.storage
-            // Two batch files are ready to be sent
-            val filePaths = listOf(
-                "/data/user/0/com.rudderstack.android.sampleapp/app_rudder-android-store/<WRITE_KEY>-0",
-                "/data/user/0/com.rudderstack.android.sampleapp/app_rudder-android-store/<WRITE_KEY>-1"
-            )
-            val fileUrlList = filePaths.joinToString(",")
-
-            // Mock storage read
-            coEvery {
-                storage.readString(StorageKeys.EVENT, String.empty())
-            } returns fileUrlList
-
-            // Mock file existence check
-            every { eventQueue.doesFileExist(any()) } returns true
-
-            val batchPayload = "test content"
-
-            // Mock messageQueue file reading
-            filePaths.forEach { path ->
-                every { eventQueue.readFileAsString(path) } returns batchPayload
-            }
-
-            // Mock the behavior for HttpClient
-            every { mockHttpClient.sendData(batchPayload) } returns Result.Success("Ok")
-
-            // Execute messageQueue actions
-            eventQueue.start()
-            eventQueue.flush()
-            testDispatcher.scheduler.advanceUntilIdle()
-
-            // Verify the expected behavior
-            filePaths.forEach { path ->
-                verify(exactly = 1) { storage.remove(path) }
-            }
-        }
-
-    @Test
-    fun `given batches of events with different anonymousIds, when they are uploaded, then header is updated for each batch with different anonymousId`() {
-        val storage = mockAnalytics.storage
-
-        val filePaths = listOf(
-            "/data/user/0/com.rudderstack.android.sampleapp/app_rudder-android-store/<WRITE_KEY>-0",
-            "/data/user/0/com.rudderstack.android.sampleapp/app_rudder-android-store/<WRITE_KEY>-1"
-        )
-        val fileUrlList = filePaths.joinToString(",")
-
-        val batchPayload1 = "test content 1"
-        val batchPayload2 = "test content 2"
-        val anonymousId1 = "anonymousId1"
-        val anonymousId2 = "anonymousId2"
-
-        coEvery {
-            storage.readString(StorageKeys.EVENT, String.empty())
-        } returns fileUrlList
-        every { eventQueue.doesFileExist(any()) } returns true
-        every { eventQueue.readFileAsString(filePaths[0]) } returns batchPayload1
-        every { eventQueue.readFileAsString(filePaths[1]) } returns batchPayload2
-
-        every { eventQueue.getAnonymousIdFromBatch(batchPayload1) } returns anonymousId1
-        every { eventQueue.getAnonymousIdFromBatch(batchPayload2) } returns anonymousId2
-
-        // Mock the behavior for HttpClient
-        every { mockHttpClient.sendData(batchPayload1) } returns Result.Success("Ok")
-        every { mockHttpClient.sendData(batchPayload2) } returns Result.Success("Ok")
-
-        // Execute messageQueue actions
-        eventQueue.start()
-        eventQueue.flush()
-
-        testDispatcher.scheduler.advanceUntilIdle()
-
-        coVerify(exactly = 1) {
-            mockHttpClient.updateAnonymousIdHeaderString(anonymousId1.encodeToBase64())
-            mockHttpClient.updateAnonymousIdHeaderString(anonymousId2.encodeToBase64())
-        }
-    }
-
-    @ParameterizedTest
-    @MethodSource("batchAnonymousIdTestProvider")
-    fun `given a batch with some anonymousId, when it is uploaded, then header is updated with correct anonymousId`(
-        batchPayload: String,
-        anonymousIdFromBatch: String
-    ) = runTest(testDispatcher) {
-        val storage = mockAnalytics.storage
-
-        val filePaths = listOf(
-            "/data/user/0/com.rudderstack.android.sampleapp/app_rudder-android-store/<WRITE_KEY>-0"
-        )
-        val fileUrlList = filePaths.joinToString(",")
-
-        coEvery {
-            storage.readString(StorageKeys.EVENT, String.empty())
-        } returns fileUrlList
-        every { eventQueue.doesFileExist(any()) } returns true
-        every { eventQueue.readFileAsString(filePaths[0]) } returns batchPayload
-
-        // Mock the behavior for HttpClient
-        every { mockHttpClient.sendData(batchPayload) } returns Result.Success("Ok")
-
-        val randomUUID = "some_random_id"
-        mockkStatic(::generateUUID)
-        every { generateUUID() } returns randomUUID
-
-        // Execute messageQueue actions
-        eventQueue.start()
-        eventQueue.flush()
-        testDispatcher.scheduler.advanceUntilIdle()
-
-        val encodedAnonymousId = anonymousIdFromBatch.encodeToBase64()
-        coVerify(atLeast = 1) {
-            mockHttpClient.updateAnonymousIdHeaderString(encodedAnonymousId)
-        }
-    }
-
-    @Test
-    fun `given batch is ready to be sent to the server and server returns error, when flush is called, then the batch is not removed from storage`() {
-        val storage = mockAnalytics.storage
-        // Two batch files are ready to be sent
-        val filePaths = listOf(
-            "/data/user/0/com.rudderstack.android.sampleapp/app_rudder-android-store/<WRITE_KEY>-0",
-            "/data/user/0/com.rudderstack.android.sampleapp/app_rudder-android-store/<WRITE_KEY>-1"
-        )
-        val fileUrlList = filePaths.joinToString(",")
-
-        // Mock storage read
-        coEvery {
-            storage.readString(StorageKeys.EVENT, String.empty())
-        } returns fileUrlList
-
-        // Mock file existence check
-        every { eventQueue.doesFileExist(any()) } returns true
-
-        val batchPayload = "test content"
-
-        // Mock messageQueue file reading
-        filePaths.forEach { path ->
-            every { eventQueue.readFileAsString(path) } returns batchPayload
-        }
-
-        // Mock the behavior for HttpClient
-        every { mockHttpClient.sendData(batchPayload) } returns Result.Failure(
-            ErrorStatus.GENERAL_ERROR,
-            IOException("Internal Server Error")
-        )
-
-        // Execute messageQueue actions
-        eventQueue.start()
-        eventQueue.flush()
-        testDispatcher.scheduler.advanceUntilIdle()
-
-        // Verify the expected behavior
-        filePaths.forEach { path ->
-            verify(exactly = 0) { storage.remove(path) }
-        }
-    }
-
-    @Test
-    fun `given batch is ready to be sent to the server and file is not found, when flush is called, then the exception is thrown and handled`() {
-        val storage = mockAnalytics.storage
-        // Two batch files are ready to be sent
-        val filePaths = listOf(
-            "/data/user/0/com.rudderstack.android.sampleapp/app_rudder-android-store/<WRITE_KEY>-0",
-            "/data/user/0/com.rudderstack.android.sampleapp/app_rudder-android-store/<WRITE_KEY>-1"
-        )
-        val fileUrlList = filePaths.joinToString(",")
-
-        // Mock storage read
-        coEvery {
-            storage.readString(StorageKeys.EVENT, String.empty())
-        } returns fileUrlList
-
-        // Mock file existence check
-        every { eventQueue.doesFileExist(any()) } returns true
-
-        // Throw file not found exception while reading the file
-        val exception = FileNotFoundException("File not found")
-        filePaths.forEach { path ->
-            every { eventQueue.readFileAsString(path) } throws exception
-        }
-
-        // Execute messageQueue actions
-        eventQueue.start()
-        eventQueue.flush()
-        testDispatcher.scheduler.advanceUntilIdle()
-
-        verify(exactly = filePaths.size) {
-            mockKotlinLogger.error("Message storage file not found", exception)
-        }
-    }
-
-    @Test
-    fun `given batch is ready to be sent to the server and some exception occurs while reading the file, when flush is called, then the exception is thrown and handled`() {
-        val storage = mockAnalytics.storage
-        // Two batch files are ready to be sent
-        val filePaths = listOf(
-            "/data/user/0/com.rudderstack.android.sampleapp/app_rudder-android-store/<WRITE_KEY>-0",
-            "/data/user/0/com.rudderstack.android.sampleapp/app_rudder-android-store/<WRITE_KEY>-1"
-        )
-        val fileUrlList = filePaths.joinToString(",")
-
-        // Mock storage read
-        coEvery {
-            storage.readString(StorageKeys.EVENT, String.empty())
-        } returns fileUrlList
-
-        // Mock file existence check
-        every { eventQueue.doesFileExist(any()) } returns true
-
-        // Throw generic exception while reading the file
-        val exception = Exception("File not found")
-        filePaths.forEach { path ->
-            every { eventQueue.readFileAsString(path) } throws exception
-        }
-
-        // Execute messageQueue actions
-        eventQueue.start()
-        eventQueue.flush()
-        testDispatcher.scheduler.advanceUntilIdle()
-
-        verify(exactly = filePaths.size) {
-            mockKotlinLogger.error("Error when uploading event", exception)
-        }
     }
 
     @Test
@@ -670,26 +414,5 @@ class EventQueueTest {
         }
     }
 
-    companion object {
 
-        @JvmStatic
-        fun batchAnonymousIdTestProvider() = listOf(
-            Arguments.of(
-                """{"userId": "12345", "anonymousId": "abc-123", "event": "test"}""",
-                "abc-123"
-            ),
-            Arguments.of(
-                """{"userId": "12345", "event": "test", "anonymousId":"xyz-456"}""",
-                "xyz-456"
-            ),
-            Arguments.of(
-                """{"anonymousId": "lmn-789"}""",
-                "lmn-789"
-            ),
-            Arguments.of(
-                """{"userId": "12345", "event": "test"}""",
-                "some_random_id"
-            )
-        )
-    }
 }
